pyproject.toml
[tool.poetry]
name = "r2e"
version = "0.5.0"
description = ""
authors = ["cesargodoi <cesar.godoi@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.x"
pillow = "^10.0.0"
django-widget-tweaks = "^1.4.12"
django-bootstrap5 = "^23.3"
dynaconf = "^3.2.3"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
ipdb = "^0.13.13"
mypy = "^1.2.0"
django-extensions = "^3.2.1"
ipython = "^8.13.2"
pytest = "^8.0.0"
pytest-django = "^4.7.0"
pytest-factoryboy = "^2.6.0"
parameterized = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
